Testing done 100%

+++ src/main/java/dev/kevinjimenez/bookv2/service/book/implementation/BookServiceImpl.java
@@ -1,6 +1,7 @@
 package dev.kevinjimenez.bookv2.service.book.implementation;
 
 import dev.kevinjimenez.bookv2.dao.book.BookDao;
+import dev.kevinjimenez.bookv2.exception.ExistingValueException;
 import dev.kevinjimenez.bookv2.exception.ValueNotFoundException;
 import dev.kevinjimenez.bookv2.model.Author;
 import dev.kevinjimenez.bookv2.model.Book;
@@ -10,6 +11,7 @@ import dev.kevinjimenez.bookv2.service.book.BookService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.validation.constraints.NotNull;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -30,10 +32,12 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public Book findById(int id) {
+    public Book findById(int id) throws ValueNotFoundException{
         BookDTO bookDTO = this.bookDao.findById(id);
-        Book book = new Book(bookDTO);
-        return book;
+        if(bookDTO == null){
+            throw new ValueNotFoundException("The book '" + id + "' doesn't exists.");
+        }
+        return new Book(bookDTO);
     }
 
     @Override
@@ -48,7 +52,10 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public void insert(Book book) {
+    public void insert(@NotNull Book book) throws ExistingValueException {
+        if (this.bookDao.findById(book.getIsbn()) != null){
+            throw new ExistingValueException("The book '" + book.getIsbn() + "' already exists.");
+        }
         BookDTO bookDTO = new BookDTO(book);
         Author author = this.authorService.findById(book.getAuthorId());
         if(author.getBooks() == null){
@@ -56,23 +63,32 @@ public class BookServiceImpl implements BookService {
         }
         author.getBooks().add(book.getIsbn());
         this.authorService.update(author);
-        this.bookDao.inset(bookDTO);
+        this.bookDao.insert(bookDTO);
     }
 
     @Override
-    public void update(Book book) throws ValueNotFoundException {
+    public void update(@NotNull Book book) throws ValueNotFoundException {
+        if (this.bookDao.findById(book.getIsbn()) == null){
+            throw new ValueNotFoundException("The book '" + book.getIsbn() + "' doesn't exist.");
+        }
         BookDTO bookDTO = new BookDTO(book);
         this.bookDao.update(bookDTO);
     }
 
     @Override
-    public void delete(Book book) {
+    public void delete(@NotNull Book book) throws ValueNotFoundException{
+        if (this.bookDao.findById(book.getIsbn()) == null){
+            throw new ValueNotFoundException("The book '" + book.getIsbn() + "' doesn't exist.");
+        }
         BookDTO bookDTO = new BookDTO(book);
         this.bookDao.delete(bookDTO);
     }
 
     @Override
-    public void delete(int id) {
+    public void delete(int id) throws ValueNotFoundException {
+        if (this.bookDao.findById(id) == null){
+            throw new ValueNotFoundException("The book '" + id + "' doesn't exist.");
+        }
         BookDTO bookDTO = this.bookDao.findById(id);
         this.bookDao.delete(bookDTO);
     }
